+++ app/src/main/java/com/and04/naturealbum/data/retorifit/ReverseGeocodeAPIFactory.kt
@@ -1,20 +1,12 @@
-package com.and04.naturealbum.di
+package com.and04.naturealbum.data.retorifit
 
 import com.and04.naturealbum.BuildConfig.NAVER_MAP_CLIENT_ID
 import com.and04.naturealbum.BuildConfig.NAVER_MAP_CLIENT_SECRET
-import com.and04.naturealbum.data.retorifit.ReverseGeocodeAPI
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
 import okhttp3.OkHttpClient
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
-import javax.inject.Singleton
 
-@Module
-@InstallIn(SingletonComponent::class)
-object ReverseGeocodeRetrofitModule {
+object ReverseGeocodeAPIFactory {
     private const val BASE_URL = "https://naveropenapi.apigw.ntruss.com/"
 
     private val client = OkHttpClient.Builder()
@@ -27,9 +19,7 @@ object ReverseGeocodeRetrofitModule {
         }
         .build()
 
-    @Provides
-    @Singleton
-    fun provideReverseGeocodeAPI(): ReverseGeocodeAPI = Retrofit.Builder()
+    fun create(): ReverseGeocodeAPI = Retrofit.Builder()
         .baseUrl(BASE_URL)
         .client(client)
         .addConverterFactory(GsonConverterFactory.create())
