+++ app/src/main/java/com/and04/naturealbum/di/ReverseGeocodeModule.kt
@@ -3,6 +3,7 @@ package com.and04.naturealbum.di
 import com.and04.naturealbum.data.repository.ReverseGeocodeRepository
 import com.and04.naturealbum.data.repository.ReverseGeocodeRepositoryImpl
 import com.and04.naturealbum.data.retorifit.ReverseGeocodeAPI
+import com.and04.naturealbum.data.retorifit.ReverseGeocodeAPIFactory
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
@@ -11,7 +12,7 @@ import javax.inject.Singleton
 
 @Module
 @InstallIn(SingletonComponent::class)
-object ReverseGeocodeRepositoryModule {
+object ReverseGeocodeModule {
 
     @Singleton
     @Provides
@@ -20,4 +21,9 @@ object ReverseGeocodeRepositoryModule {
     ): ReverseGeocodeRepository {
         return ReverseGeocodeRepositoryImpl(reverseGeocodeAPI)
     }
+
+    @Provides
+    @Singleton
+    fun provideReverseGeocodeAPI(): ReverseGeocodeAPI =
+        ReverseGeocodeAPIFactory.create()
 }
